+++ chalice/app.py
@@ -1,6 +1,9 @@
 """Chalice app and routing code."""
 import re
+import sys
 import base64
+import logging
+from collections import Mapping
 
 # Implementation note:  This file is intended to be a standalone file
 # that gets copied into the lambda deployment package.  It has no dependencies
@@ -39,6 +42,10 @@ class NotFoundError(ChaliceViewError):
     STATUS_CODE = 404
 
 
+class MethodNotAllowedError(ChaliceViewError):
+    STATUS_CODE = 405
+
+
 class ConflictError(ChaliceViewError):
     STATUS_CODE = 409
 
@@ -57,17 +64,42 @@ ALL_ERRORS = [
     TooManyRequestsError]
 
 
+class CaseInsensitiveMapping(Mapping):
+    """Case insensitive and read-only mapping."""
+
+    def __init__(self, mapping):
+        self._dict = {k.lower(): v for k, v in mapping.items()}
+
+    def __getitem__(self, key):
+        return self._dict[key.lower()]
+
+    def __iter__(self):
+        return iter(self._dict)
+
+    def __len__(self):
+        return len(self._dict)
+
+    def __repr__(self):
+        return 'CaseInsensitiveMapping(%s)' % repr(self._dict)
+
+
 class Request(object):
     """The current request from API gateway."""
 
     def __init__(self, query_params, headers, uri_params, method, body,
-                 base64_body, context, stage_vars):
+                 base64_body, context, claims, stage_vars):
         self.query_params = query_params
-        self.headers = headers
+        self.headers = CaseInsensitiveMapping(headers)
         self.uri_params = uri_params
         self.method = method
-        #: The parsed JSON from the body.
-        self.json_body = body
+        self.claims = claims
+        #: The parsed JSON from the body.  This value should
+        #: only be set if the Content-Type header is application/json,
+        #: which is the default content type value in chalice.
+        if self.headers.get('content-type') == 'application/json':
+            self.json_body = body
+        else:
+            self.json_body = None
         # This is the raw base64 body.
         # We'll only bother decoding this if the user
         # actually requests this via the `.raw_body` property.
@@ -84,20 +116,19 @@ class Request(object):
         return self._raw_body
 
     def to_dict(self):
-        return self.__dict__.copy()
+        copied = self.__dict__.copy()
+        # We want the output of `to_dict()` to be
+        # JSON serializable, so we need to remove the CaseInsensitive dict.
+        copied['headers'] = dict(copied['headers'])
+        return copied
 
 
 class RouteEntry(object):
 
-    def __init__(
-            self,
-            view_function,
-            view_name,
-            path,
-            methods,
-            authorization_type=None,
-            authorizer_id=None,
-            api_key_required=False):
+    def __init__(self, view_function, view_name, path, methods,
+                 authorization_type=None, authorizer_id=None,
+                 api_key_required=None, content_types=None,
+                 cors=False):
         self.view_function = view_function
         self.view_name = view_name
         self.uri_pattern = path
@@ -108,6 +139,8 @@ class RouteEntry(object):
         #: A list of names to extract from path:
         #: e.g, '/foo/{bar}/{baz}/qux -> ['bar', 'baz']
         self.view_args = self._parse_view_args()
+        self.content_types = content_types
+        self.cors = cors
 
     def _parse_view_args(self):
         if '{' not in self.uri_pattern:
@@ -118,21 +151,48 @@ class RouteEntry(object):
         return results
 
     def __eq__(self, other):
-        return (
-            self.view_function == other.view_function and
-            self.view_name == other.view_name and
-            self.uri_pattern == other.uri_pattern and
-            self.view_args == other.view_args
-        )
+        return self.__dict__ == other.__dict__
 
 
 class Chalice(object):
 
-    def __init__(self, app_name):
+    FORMAT_STRING = '%(name)s - %(levelname)s - %(message)s'
+
+    def __init__(self, app_name, configure_logs=True):
         self.app_name = app_name
         self.routes = {}
         self.current_request = None
         self.debug = False
+        self.configure_logs = configure_logs
+        self.log = logging.getLogger(self.app_name)
+        if self.configure_logs:
+            self._configure_logging()
+
+    def _configure_logging(self):
+        log = logging.getLogger(self.app_name)
+        if self._already_configured(log):
+            return
+        handler = logging.StreamHandler(sys.stdout)
+        # Timestamp is handled by lambda itself so the
+        # default FORMAT_STRING doesn't need to include it.
+        formatter = logging.Formatter(self.FORMAT_STRING)
+        handler.setFormatter(formatter)
+        log.propagate = False
+        if self.debug:
+            level = logging.DEBUG
+        else:
+            level = logging.ERROR
+        log.setLevel(level)
+        log.addHandler(handler)
+
+    def _already_configured(self, log):
+        if not log.handlers:
+            return False
+        for handler in log.handlers:
+            if isinstance(handler, logging.StreamHandler):
+                if handler.stream == sys.stdout:
+                    return True
+        return False
 
     def route(self, path, **kwargs):
         def _register_view(view_func):
@@ -141,24 +201,29 @@ class Chalice(object):
         return _register_view
 
     def _add_route(self, path, view_func, **kwargs):
-        name = kwargs.get('name', view_func.__name__)
-        methods = kwargs.get('methods', ['GET'])
-        authorization_type = kwargs.get('authorization_type', None)
-        authorizer_id = kwargs.get('authorizer_id', None)
-        api_key_required = kwargs.get('api_key_required', None)
+        name = kwargs.pop('name', view_func.__name__)
+        methods = kwargs.pop('methods', ['GET'])
+        authorization_type = kwargs.pop('authorization_type', None)
+        authorizer_id = kwargs.pop('authorizer_id', None)
+        api_key_required = kwargs.pop('api_key_required', None)
+        content_types = kwargs.pop('content_types', ['application/json'])
+        cors = kwargs.pop('cors', False)
+        if not isinstance(content_types, list):
+            raise ValueError('In view function "%s", the content_types '
+                             'value must be a list, not %s: %s'
+                             % (name, type(content_types), content_types))
+        if kwargs:
+            raise TypeError('TypeError: route() got unexpected keyword '
+                            'arguments: %s' % ', '.join(list(kwargs)))
 
         if path in self.routes:
             raise ValueError(
                 "Duplicate route detected: '%s'\n"
                 "URL paths must be unique." % path)
-        self.routes[path] = RouteEntry(
-            view_func,
-            name,
-            path,
-            methods,
-            authorization_type,
-            authorizer_id,
-            api_key_required)
+        entry = RouteEntry(view_func, name, path, methods, authorization_type,
+                           authorizer_id, api_key_required,
+                           content_types, cors)
+        self.routes[path] = entry
 
     def __call__(self, event, context):
         # This is what's invoked via lambda.
@@ -175,7 +240,7 @@ class Chalice(object):
             raise ChaliceError("No view function for: %s" % resource_path)
         route_entry = self.routes[resource_path]
         if http_method not in route_entry.methods:
-            raise ChaliceError("Unsupported method: %s" % http_method)
+            raise MethodNotAllowedError("Unsupported method: %s" % http_method)
         view_function = route_entry.view_function
         function_args = [event['params']['path'][name]
                          for name in route_entry.view_args]
@@ -187,9 +252,14 @@ class Chalice(object):
                                        event['body-json'],
                                        event['base64-body'],
                                        event['context'],
+                                       event['claims'],
                                        event['stage-variables'])
         try:
             response = view_function(*function_args)
+        except ChaliceViewError:
+            # Any chalice view error should propagate.  These
+            # get mapped to various HTTP status codes in API Gateway.
+            raise
         except Exception:
             if self.debug:
                 # If the user has turned on debug mode,