+++ tests/unit/test_analyzer.py
@@ -21,18 +21,18 @@ def chalice_aws_calls(source_code):
 
 def known_types_for_module(source_code):
     real_source_code = dedent(source_code)
-    t = analyzer.SymbolTableTypeInfer()
     compiled = analyzer.parse_code(real_source_code)
-    t.bind_types(compiled)
+    t = analyzer.SymbolTableTypeInfer(compiled)
+    t.bind_types()
     known = t.known_types()
     return known
 
 
 def known_types_for_function(source_code, name):
     real_source_code = dedent(source_code)
-    t = analyzer.SymbolTableTypeInfer()
     compiled = analyzer.parse_code(real_source_code)
-    t.bind_types(compiled)
+    t = analyzer.SymbolTableTypeInfer(compiled)
+    t.bind_types()
     known = t.known_types(scope_name=name)
     return known
 
@@ -417,6 +417,22 @@ def test_can_handle_lambda_keyword():
         foo(12)
     """) == {}
 
+
+def test_dict_comp_with_no_client_calls():
+    assert aws_calls("""\
+        import boto3
+        foo = {i: i for i in range(10)}
+    """) == {}
+
+
+#def test_can_handle_dict_comp():
+#    assert aws_calls("""\
+#        import boto3
+#        ddb = boto3.client('dynamodb')
+#        tables = {t: t for t in ddb.list_tables()}
+#    """) == {'dynamodb': set(['list_tables'])}
+#
+#
 #def test_tuple_assignment():
 #    assert aws_calls("""\
 #        import boto3